+++ mission-racing/src/main/java/com/racing/controller/WebController.java
@@ -1,0 +1,25 @@
+package com.racing.controller;
+
+import com.racing.domain.Cars;
+import com.racing.domain.dto.StartRaceRequest;
+import com.racing.domain.dto.WinnerResponse;
+import com.racing.service.RacingService;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.web.bind.annotation.*;
+
+@RestController
+@RequestMapping("/api/cars")
+public class WebController {
+
+    private final RacingService racingService;
+
+    public WebController(RacingService racingService) {
+        this.racingService = racingService;
+    }
+
+    @PostMapping("/racing")
+    public WinnerResponse startRace(@RequestBody StartRaceRequest request) {
+        Cars cars = racingService.createCars(request.getCarNames());
+        return racingService.race(cars, request.getTryCount());
+    }
+}
