+++ mission-racing/src/main/java/com/racing/domain/RandomNumberGenerate.java
@@ -1,0 +1,11 @@
+package com.racing.domain;
+
+import org.springframework.stereotype.Component;
+
+@Component
+public class RandomNumberGenerate implements NumberGenerator {
+    @Override
+    public int generateRandom() {
+        return (int) (Math.random() * 10);
+    }
+}
