+++ mission-racing/src/main/java/com/racing/domain/dto/StartRaceRequest.java
@@ -1,0 +1,15 @@
+package com.racing.domain.dto;
+
+public class StartRaceRequest {
+
+    private String carNames;
+    private int tryCount;
+
+    public String getCarNames() {
+        return carNames;
+    }
+
+    public int getTryCount() {
+        return tryCount;
+    }
+}
