+++ mission-racing/src/main/java/com/racing/domain/dto/WinnerResponse.java
@@ -1,0 +1,27 @@
+package com.racing.domain.dto;
+
+import org.springframework.web.bind.annotation.GetMapping;
+
+import java.util.List;
+
+public class WinnerResponse {
+
+    private final List<String> carStates;
+    private final List<String> winners;
+
+    public WinnerResponse(final List<String> carStates, final List<String> winners) {
+        this.carStates = carStates;
+        this.winners = winners;
+    }
+
+    @GetMapping("api/cars/result")
+    public void printResultDto() {
+        for (String carState : carStates) {
+            System.out.println(carState);
+        }
+    }
+
+    public List<String> getWinners() {
+        return winners;
+    }
+}
