+++ mission-racing/src/main/java/com/racing/domain/vo/Name.java
@@ -1,0 +1,27 @@
+package com.racing.domain.vo;
+
+import com.racing.view.exception.NullInputException;
+
+public class Name {
+
+    private final String name;
+
+    public static Name from(final String name) {
+        validateIsBlank(name);
+        return new Name(name);
+    }
+
+    private Name(final String carName) {
+        this.name = carName;
+    }
+
+    private static void validateIsBlank(final String name) {
+        if (name.isBlank()) {
+            throw new NullInputException();
+        }
+    }
+
+    public String getName() {
+        return name;
+    }
+}
