+++ mission-racing/src/main/java/com/racing/view/exception/NullInputException.java
@@ -1,0 +1,8 @@
+package com.racing.view.exception;
+
+public class NullInputException extends RuntimeException {
+
+    public NullInputException() {
+        super("자동차 이름을 꼭 입력하세요");
+    }
+}
