--- mission-racing/src/racingSpring/src/main/java/com/example/racingspring/ConsoleApplication.java
@@ -1,21 +1,0 @@
-package com.example.racingspring;
-
-
-import com.example.racingspring.controller.RacingController;
-import com.example.racingspring.domain.RandomNumberGenerate;
-import com.example.racingspring.view.Inputview;
-import com.example.racingspring.view.Outputview;
-
-public class ConsoleApplication {
-
-    public static void main(String[] args) {
-        Inputview inputview = new Inputview();
-        Outputview outputview = new Outputview();
-        RandomNumberGenerate numberGenerator = new RandomNumberGenerate();
-        RacingController racingController = new RacingController(
-                inputview,
-                outputview,
-                numberGenerator);
-        racingController.run();
-    }
-}
