--- mission-racing/src/racingSpring/src/main/java/com/example/racingspring/domain/Car.java
@@ -1,31 +1,0 @@
-package com.example.racingspring.domain;
-
-import com.example.racingspring.domain.vo.MoveCount;
-import com.example.racingspring.domain.vo.Name;
-
-public class Car {
-
-    private final Name name;
-    private final MoveCount moveCount;
-
-    public Car(final String name) {
-        this.name = Name.from(name);
-        this.moveCount = new MoveCount();
-    }
-
-    public boolean isSameCount(final int count) {
-        return this.moveCount.isSame(count);
-    }
-
-    public void move(final int randomNumber) {
-        this.moveCount.move(randomNumber);
-    }
-
-    public int getMoveCount() {
-        return moveCount.getMoveCount();
-    }
-
-    public String getName() {
-        return name.getName();
-    }
-}
