--- mission-racing/src/racingSpring/src/main/java/com/example/racingspring/domain/NumberGenerator.java
@@ -1,5 +1,0 @@
-package com.example.racingspring.domain;
-
-public interface NumberGenerator {
-    int generateRandom();
-}
