--- mission-racing/src/racingSpring/src/main/java/com/example/racingspring/domain/RandomNumberGenerate.java
@@ -1,11 +1,0 @@
-package com.example.racingspring.domain;
-
-import org.springframework.stereotype.Component;
-
-@Component
-public class RandomNumberGenerate implements NumberGenerator {
-    @Override
-    public int generateRandom() {
-        return (int) (Math.random() * 10);
-    }
-}
