--- mission-racing/src/racingSpring/src/main/java/com/example/racingspring/domain/dto/StartRaceRequest.java
@@ -1,15 +1,0 @@
-package com.example.racingspring.domain.dto;
-
-public class StartRaceRequest {
-
-    private String carNames;
-    private int tryCount;
-
-    public String getCarNames() {
-        return carNames;
-    }
-
-    public int getTryCount() {
-        return tryCount;
-    }
-}
