--- mission-racing/src/racingSpring/src/main/java/com/example/racingspring/domain/도메인로직.md
@@ -1,79 +1,0 @@
-## 요구 사항 분석
-
-- 자동차는 전진 혹은 멈출 수 있다.
-- 자동차는 이름이 있다.
-    - 자동차의 이름은 5자 이하이다.
-    - 자동차의 이름은 공백이 불가하고 콤마로 구분한다.
-    - 자동차는 n번 이동이 가능하다.
-    - 자동차의 전진은 0~9 랜덤 값중 4이상일 경우이고, 그 외에는 멈춘다.
-- 자동차 경주 게임의 라운드 한 번당 중간 결과를 출력해야한다.
-- 자동차 경주 게임의 우승자는 한 명 이상일 수 있다.
-
-
-
-
-
-----------
-domain
-- [x] 자동차 일급컬렉션
-- [x] 자동차 생성
-- [x] 랜덤 넘버 뽑기
-- [x] 랜덤 넘버로 움직이게하기
-- [x] 승자 구하기 
-
-
-경주할 자동차 이름을 입력하세요.(이름은 쉼표(,) 기준으로 구분)
-pobi,woni,jun
-시도할 회수는 몇회인가요?
-5
-
-실행 결과
-pobi : -
-woni :
-jun : -
-
-pobi : --
-woni : -
-jun : --
-
-pobi : ---
-woni : --
-jun : ---
-
-pobi : ----
-woni : ---
-jun : ----
-
-pobi : -----
-woni : ----
-jun : -----
-
-최종 우승자 : pobi, jun
-
-
-
-
-
----------------------------------------------------------------------
-Spring 미션 
-
-
----------------------------------------------------------------------
-Controller 
-- 웹 요청처리를한다. 
-
-
-
-
----------------------------------------------------------------------
-Service 
-비즈니스 로직을 처리한다.  
-즉 도메인에서 최종 도출해내야할 것 - > 우승자 
-우승자를 서비스레이어에서 처리하게해야겠다. 
-
-
----------------------------------------------------------------------
-Repository 
-
-데이터베이스와 상호작용을 한다 
-여기서는 자동차를 입력하는 데이터를 데이터베이스에 보내야 할 것 같다.
