--- mission-racing/src/racingSpring/src/main/java/com/example/racingspring/view/exception/NullInputException.java
@@ -1,8 +1,0 @@
-package com.example.racingspring.view.exception;
-
-public class NullInputException extends RuntimeException {
-
-    public NullInputException() {
-        super("자동차 이름을 꼭 입력하세요");
-    }
-}
